+++ fullTestManualStand_TestWindow.cpp
@@ -35,7 +35,7 @@ void TestWindow::initUiTableFullTestManualStand()
 void TestWindow::resetTableHeaderFullTestManualStand()
 {
 
-	mainTableWidget->setRowCount(cableRows.size());
+	mainTableWidget->setRowCount((int)cableRows.size());
 	mainTableWidget->setColumnCount(COLUMN_COUNT);
 	
 	resetTableHeaderLanguageFullTestManualStand();
