+++ inTestManualStand_TestWindow.cpp
@@ -25,7 +25,7 @@ void TestWindow::initUiTableInTestManualStand()
 
 void TestWindow::initUiTableHeaderInTestManualStand()
 {
-	mainTableWidget->setRowCount(cableRows.size());
+	mainTableWidget->setRowCount((int)cableRows.size());
 	mainTableWidget->setColumnCount(COLUMN_COUNT);
 
 	resetTableHeaderLanguageInTestManualStand();
